feat(blog-client): :art: modified features with notion and log
