+++ apps/blog-client/app/[locale]/surflog/[slug]/page.tsx
@@ -1,8 +1,7 @@
 // prismjs
-import { LogDetailRenderer } from '@/features'
+import { getSurflogDetail, LogDetailRenderer, queryLogs } from '@/features'
 import { queryKeyFactory } from '@/lib/react-query/react-query.key-factory'
 import { getQueryClient } from '@/lib/react-query/react-query.utils'
-import { getSurflogDetail, queryLogs } from '@/lib/utils'
 import { dehydrate, HydrationBoundary } from '@tanstack/react-query'
 import { routing } from 'i18n/routing'
 import { PageProps } from 'i18n/types'
