+++ apps/blog-client/app/[locale]/techlog/[slug]/page.tsx
@@ -1,7 +1,6 @@
-import { LogDetailRenderer } from '@/features'
+import { getTechlogDetail, LogDetailRenderer, queryLogs } from '@/features'
 import { queryKeyFactory } from '@/lib/react-query/react-query.key-factory'
 import { getQueryClient } from '@/lib/react-query/react-query.utils'
-import { getTechlogDetail, queryLogs } from '@/lib/utils'
 import { dehydrate, HydrationBoundary } from '@tanstack/react-query'
 import { routing } from 'i18n/routing'
 import { PageProps } from 'i18n/types'
