+++ apps/blog-client/app/api/resume/route.ts
@@ -1,4 +1,4 @@
-import { queryNotionResumePage } from '@/lib'
+import { queryNotionResumePage } from '@/features/notion'
 import { getBlocks } from '@/lib/notion'
 import { NextRequest, NextResponse } from 'next/server'
