+++ apps/blog-client/features/logs/types/index.ts
@@ -1,0 +1,1 @@
+export * from './platform'
