+++ apps/blog-client/features/logs/types/platform.ts
@@ -1,0 +1,1 @@
+export type LogPlatform = 'techlog' | 'surflog' | 'filmlog' | 'soundlog' | 'squarelog' | 'textlog'
