+++ apps/blog-client/features/notion/index.js
@@ -1,1 +1,2 @@
+export * from './query'
 export * from './renderer'
