+++ apps/blog-client/features/pdf/pdf.utils.ts
@@ -1,0 +1,24 @@
+export const generatePDF = async () => {
+  // Set options for html2pdf
+  const options = {
+    margin: 0.25,
+    filename: 'website_screenshot.pdf',
+    image: { type: 'jpeg', quality: 0.98 },
+    html2canvas: { scale: 1.5 },
+    jsPDF: { unit: 'in', format: 'letter', orientation: 'portrait' },
+    pagebreak: { mode: 'avoid-all' },
+  }
+
+  // Select the element to capture
+  const element = document.body
+
+  // @ts-expect-error
+  const { default: html2pdf } = await import('html2pdf.js')
+  // Generate PDF
+  html2pdf()
+    .from(element)
+    .set({
+      ...options,
+    })
+    .save()
+}
