+++ apps/blog-client/lib/fetchers/index.ts
@@ -1,6 +1,6 @@
+import { queryLogs } from '@/features/notion'
 import { PageObjectResponse, PersonUserObjectResponse } from '@notionhq/client/build/src/api-endpoints'
 import { AppLocale } from 'i18n/types'
-import { queryLogs } from '../utils'
 
 const BASE_URL = process.env.NODE_ENV === 'development' ? 'http://localhost:3000' : 'https://blog.coldsurf.io'
