+++ apps/blog-client/lib/index.ts
@@ -1,3 +1,2 @@
 export * from './react-query'
 export * from './registries'
-export * from './utils'
